PUT api for add hobby to person works

+++ src/main/java/facades/IHobbyFacade.java
@@ -9,7 +9,7 @@ public interface IHobbyFacade {
 
     public List<HobbyDTO> getAllHobbies();
 
-    public HobbyDTO getHobbyById(int hobbyId);
+    public HobbyDTO getHobbyById(Long hobbyId);
 
     //US2
     public List<PersonDTO> getPersonsByHobby(HobbyDTO hobby);
